refactor and first unittests
