+++ libs/chain-signatures/contract/src/state/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod test_key_state;
+pub mod test_utils;
